src/requirements.txt
chainlit
requests
openai
python-dotenv
